Improve uom support.
Include an extensive list of mappings, including all of the Energistics
units in witsmlUnitDict.xml. The strategy is to map any unit encountered
in a LAS file by first lowercasing it, then looking it up in the uom
mapping table. Support using an external mapping file with the -m option.